refactor: move `FunCommandSet` to root package
